move firestore module in auth.js
