+++ res/scripts/workplace.js
@@ -13,44 +13,3 @@ monitorAuthState((user) => {
     window.location.href = "index.html";
   }
 })
-
-import { db, setDoc, doc } from '../scripts/firebase/firebase-config.js';
-
-    function getAccessTokenFromUrl() {
-      const hash = window.location.hash;
-      const params = new URLSearchParams(hash.substring(1));
-      return params.get('access_token');
-    }
-    
-    const accessToken = getAccessTokenFromUrl();
-    if (accessToken) {
-      fetchUserInfo(accessToken);
-    }
-
-    async function fetchUserInfo(token) {
-      try {
-        const response = await fetch('https://www.googleapis.com/oauth2/v3/userinfo', {
-          headers: { 'Authorization': `Bearer ${token}` }
-        });
-        const userInfo = await response.json();
-
-        document.getElementById('profileImage').src = userInfo.picture;
-        document.getElementById('profileImage').alt = userInfo.name;
-        document.getElementById('profileCardImage').src = userInfo.picture;
-        document.getElementById('profileCardImage').alt = userInfo.name;
-        document.getElementById('profileCardName').textContent = `Hi, ${userInfo.name}!`;
-        document.getElementById('profileCardEmail').textContent = userInfo.email;
-
-        await setDoc(doc(db, "users", userInfo.sub), {
-          uid: userInfo.sub,
-          name: userInfo.name,
-          email: userInfo.email,
-          picture: userInfo.picture
-        }, { merge: true });
-
-        console.log('User information stored in Firestore.');
-      } catch (error) {
-        console.error('Error fetching user info or storing in Firestore:', error);
-      }
-    }
-
